Move timeseries chart into a separate directory so it can be used as a plugable widget
